xorBlock on 386 arch
The xorBlock function expects the wordSize to be 8 bytes. On arch 386
this size is 4 bytes and xorBlock will panic. xorBlock is moved to
hash_amd64.go and hash_386.go.
